refactor: conversão de double para float para economizar memória ao processar

+++ src/main/java/com/pht/rinha_backend_2026/dto/Customer.java
@@ -1,4 +1,4 @@
 package com.pht.rinha_backend_2026.dto;
 
-public record Customer(double avg_amount, int tx_count_24h, String[] known_merchants) {
+public record Customer(float avg_amount, int tx_count_24h, String[] known_merchants) {
 }
