+++ src/main/java/com/pht/rinha_backend_2026/dto/Merchant.java
@@ -1,4 +1,4 @@
 package com.pht.rinha_backend_2026.dto;
 
-public record Merchant(String id, String mcc, double avg_amount) {
+public record Merchant(String id, String mcc, float avg_amount) {
 }
