+++ src/main/java/com/pht/rinha_backend_2026/dto/Terminal.java
@@ -1,4 +1,4 @@
 package com.pht.rinha_backend_2026.dto;
 
-public record Terminal(boolean is_online, boolean card_present, double km_from_home) {
+public record Terminal(boolean is_online, boolean card_present, float km_from_home) {
 }
