+++ src/main/java/com/pht/rinha_backend_2026/dto/Transaction.java
@@ -2,5 +2,5 @@ package com.pht.rinha_backend_2026.dto;
 
 import java.time.Instant;
 
-public record Transaction(double amount, int installments, Instant requested_at) {
+public record Transaction(float amount, int installments, Instant requested_at) {
 }
